+++ UniversityAdmission/Class/Applicant.cs
@@ -1,0 +1,32 @@
+namespace UniversityAdmission.Domain.Class;
+
+/// <summary>
+/// Абитуриент
+/// </summary>]
+public class Applicant
+{
+    /// <summary>
+    /// Идентификатор абитуриента
+    /// </summary>
+    public required int IdApplicant { get; set; }
+
+    /// <summary>
+    /// ФИО абитуриента
+    /// </summary>
+    public required string FullName { get; set; }
+
+    /// <summary>
+    /// Дата рождения абитуриента
+    /// </summary>
+    public required DateOnly DateOfBirth { get; set; }
+
+    /// <summary>
+    /// Страна абитуриента
+    /// </summary>
+    public required string Country { get; set; }
+
+    /// <summary>
+    /// Город абитуриента
+    /// </summary>
+    public required string City { get; set; }
+}
